logging: refactorings, pt. 2: error handling
fix: no errors during umount_chroot()
no progress for CI=true builds
WiP disable kernel hashing. too crazy.
a few builds now working with "set -e"
wtf. or something
kernel logging and long_running stuff - a mess - needs moving around in files
rewrite uboot compile loop without using subshells. remove ALL traps. refactor host command
better logging for u-boot
more fixes, u-boot
more fixes for logging et al
git stuff
many (many) fixes
new color scheme
a monster. make sure unmounted at the end. remove set -e's, to-be-readded.
remove set -e "for safety reasons"
more alerts. we gotta know whats failing
some more logging stuff and fixes for error checking
more logging and error handling stuff
fixes; some set -e's
more logging stuff

+++ lib/functions/misc_image.sh
@@ -28,48 +28,91 @@ mount_chroot() {
 # helper to reduce code duplication
 #
 umount_chroot() {
-
 	local target=$1
 	display_alert "Unmounting" "$target" "info"
 	while grep -Eq "${target}.*(dev|proc|sys)" /proc/mounts; do
-		umount -l --recursive "${target}"/dev > /dev/null 2>&1
-		umount -l "${target}"/proc > /dev/null 2>&1
-		umount -l "${target}"/sys > /dev/null 2>&1
-		sleep 5
+		umount --recursive "${target}"/dev > /dev/null 2>&1 || true
+		umount "${target}"/proc > /dev/null 2>&1 || true
+		umount "${target}"/sys > /dev/null 2>&1 || true
+		sync
+	done
+}
+
+# demented recursive version, for final umount.
+umount_chroot_recursive() {
+	set +e # really, ignore errors. we wanna unmount everything and will try very hard.
+	local target="$1"
+
+	if [[ ! -d "${target}" ]]; then # only even try if target is a directory
+		return 0                       # success, nothing to do.
+	fi
+	display_alert "Unmounting recursively" "$target" ""
+	sync # sync. coalesce I/O. wait for writes to flush to disk. it might take a second.
+	# First, try to umount some well-known dirs, in a certain order. for speed.
+	local -a well_known_list=("dev/pts" "dev" "proc" "sys" "boot/efi" "boot/firmware" "boot" "tmp" ".")
+	for well_known in "${well_known_list[@]}"; do
+		umount --recursive "${target}${well_known}" &> /dev/null && sync
 	done
 
+	# now try in a loop to unmount all that's still mounted under the target
+	local -i tries=1 # the first try above
+	mapfile -t current_mount_list < <(cut -d " " -f 2 "/proc/mounts" | grep "^${target}")
+	while [[ ${#current_mount_list[@]} -gt 0 ]]; do
+		if [[ $tries -gt 10 ]]; then
+			display_alert "${#current_mount_list[@]} dirs still mounted after ${tries} tries:" "${current_mount_list[*]}" "wrn"
+		fi
+		cut -d " " -f 2 "/proc/mounts" | grep "^${target}" | xargs -n1 umount --recursive &> /dev/null
+		sync # wait for fsync, then count again for next loop.
+		mapfile -t current_mount_list < <(cut -d " " -f 2 "/proc/mounts" | grep "^${target}")
+		tries=$((tries + 1))
+	done
+
+	display_alert "Unmounted OK after ${tries} attempt(s)" "$target" "info"
+	return 0
 }
 
 # unmount_on_exit
 #
 unmount_on_exit() {
+	trap - ERR           # Also remove any error trap. it's too late for that.
+	set +e               # we just wanna plow through this, ignoring errors.
 	trap - INT TERM EXIT # remove the trap
-	local stacktrace="$(get_extension_hook_stracktrace "${BASH_SOURCE[*]}" "${BASH_LINENO[*]}")"
+
+	local stacktrace
+	stacktrace="$(get_extension_hook_stracktrace "${BASH_SOURCE[*]}" "${BASH_LINENO[*]}")"
+	display_alert "trap caught, shutting down" "${stacktrace}" "err"
 	if [[ "${ERROR_DEBUG_SHELL}" == "yes" ]]; then
 		ERROR_DEBUG_SHELL=no # dont do it twice
 		display_alert "MOUNT" "${MOUNT}" "err"
 		display_alert "SDCARD" "${SDCARD}" "err"
 		display_alert "ERROR_DEBUG_SHELL=yes, starting a shell." "ERROR_DEBUG_SHELL" "err"
-		bash < /dev/tty || true
+		bash < /dev/tty >&2 || true
 	fi
 
-	umount_chroot "${SDCARD}/"
-	mountpoint -q "${SRC}"/cache/toolchain && umount -l "${SRC}"/cache/toolchain
-	mountpoint -q "${SRC}"/cache/rootfs && umount -l "${SRC}"/cache/rootfs
-	umount -l "${SDCARD}"/tmp > /dev/null 2>&1
-	umount -l "${SDCARD}" > /dev/null 2>&1
-	umount -l "${MOUNT}"/boot > /dev/null 2>&1
-	umount -l "${MOUNT}" > /dev/null 2>&1
-	[[ $CRYPTROOT_ENABLE == yes ]] && cryptsetup luksClose "${ROOT_MAPPER}"
-	losetup -d "${LOOP}" > /dev/null 2>&1
-	rm -rf --one-file-system "${SDCARD}"
+	cd "${SRC}" || echo "Failed to cwd to ${SRC}" # Move pwd away, so unmounts work
+	# those will loop until they're unmounted.
+	umount_chroot_recursive "${SDCARD}/"
+	umount_chroot_recursive "${MOUNT}/"
 
-	# if we've been called by exit_with_error itself, don't recurse. it makes no sense.
+	mountpoint -q "${SRC}"/cache/toolchain && umount -l "${SRC}"/cache/toolchain >&2 # @TODO: why does Igor uses lazy umounts? nfs?
+	mountpoint -q "${SRC}"/cache/rootfs && umount -l "${SRC}"/cache/rootfs >&2
+	[[ $CRYPTROOT_ENABLE == yes ]] && cryptsetup luksClose "${ROOT_MAPPER}" >&2
+
+	# shellcheck disable=SC2153 # global var. also a local 'loop' in another function. sorry.
+	if [[ -b "${LOOP}" ]]; then
+		display_alert "Freeing loop" "unmount_on_exit ${LOOP}" "wrn"
+		losetup -d "${LOOP}" >&2
+	fi
+
+	[[ -d "${SDCARD}" ]] && rm -rf --one-file-system "${SDCARD}"
+	[[ -d "${MOUNT}" ]] && rm -rf --one-file-system "${MOUNT}"
+
+	# if we've been called by exit_with_error itself, don't recurse.
 	if [[ "${ALREADY_EXITING_WITH_ERROR:-no}" != "yes" ]]; then
 		exit_with_error "generic error during build_rootfs_image: ${stacktrace}" || true # but don't trigger error again
 	fi
 
-	return 0 # exit successfully. we're already handling a trap here.
+	return 47 # trap returns error. # exit successfully. we're already handling a trap here.
 }
 
 # check_loop_device <device_node>
@@ -94,23 +137,33 @@ check_loop_device() {
 write_uboot() {
 
 	local loop=$1 revision
-	display_alert "Writing U-boot bootloader" "$loop" "info"
+	display_alert "Preparing u-boot bootloader" "$loop" "info"
 	TEMP_DIR=$(mktemp -d || exit 1)
 	chmod 700 ${TEMP_DIR}
 	revision=${REVISION}
 	if [[ -n $UPSTREM_VER ]]; then
 		revision=${UPSTREM_VER}
-		dpkg -x "${DEB_STORAGE}/linux-u-boot-${BOARD}-${BRANCH}_${revision}_${ARCH}.deb" ${TEMP_DIR}/
+		dpkg -x "${DEB_STORAGE}/linux-u-boot-${BOARD}-${BRANCH}_${revision}_${ARCH}.deb" ${TEMP_DIR}/ 2>&1
 	else
-		dpkg -x "${DEB_STORAGE}/${CHOSEN_UBOOT}_${revision}_${ARCH}.deb" ${TEMP_DIR}/
+		dpkg -x "${DEB_STORAGE}/${CHOSEN_UBOOT}_${revision}_${ARCH}.deb" ${TEMP_DIR}/ 2>&1
 	fi
 
-	# source platform install to read $DIR
-	source ${TEMP_DIR}/usr/lib/u-boot/platform_install.sh
-	write_uboot_platform "${TEMP_DIR}${DIR}" "$loop"
-	[[ $? -ne 0 ]] && exit_with_error "U-boot bootloader failed to install" "@host"
+	if [[ ! -f "${TEMP_DIR}/usr/lib/u-boot/platform_install.sh" ]]; then
+		exit_with_error "Missing ${TEMP_DIR}/usr/lib/u-boot/platform_install.sh"
+	fi
+
+	display_alert "Sourcing u-boot install functions" "$loop" "info"
+	source ${TEMP_DIR}/usr/lib/u-boot/platform_install.sh 2>&1
+
+	display_alert "Writing u-boot bootloader" "$loop" "info"
+	write_uboot_platform "${TEMP_DIR}${DIR}" "$loop" 2>&1
+	[[ $? -ne 0 ]] && {
+		rm -rf ${TEMP_DIR}
+		exit_with_error "U-boot bootloader failed to install" "@host"
+	}
 	rm -rf ${TEMP_DIR}
 
+	return 0
 }
 
 # copy_all_packages_files_for <folder> to package
@@ -124,6 +177,12 @@ copy_all_packages_files_for() {
 			display_alert "Adding files from" "${package_dirpath}"
 		fi
 	done
+}
+
+apt_purge_unneeded_packages() {
+	# remove packages that are no longer needed. rootfs cache + uninstall might have leftovers.
+	display_alert "No longer needed packages" "purge" "info"
+	chroot_sdcard_apt_get autoremove
 }
 
 customize_image() {
@@ -177,15 +236,69 @@ function chroot_sdcard_apt_get() {
 
 # please, please, unify around this function. if SDCARD is not enough, I'll make a mount version.
 function chroot_sdcard() {
+	run_host_command_logged_raw chroot "${SDCARD}" /bin/bash -e -c "$*"
+}
+
+function chroot_custom_long_running() {
+	local target=$1
+	shift
+	local _exit_code=1
+	if [[ "${SHOW_LOG}" == "yes" ]] || [[ "${CI}" == "true" ]]; then
+		run_host_command_logged_raw chroot "${target}" /bin/bash -e -c "$*"
+		_exit_code=$?
+	else
+		run_host_command_logged_raw chroot "${target}" /bin/bash -e -c "$*" | pv -N "$(logging_echo_prefix_for_pv "${INDICATOR:-compile}")" --progress --timer --line-mode --force --cursor --delay-start 0 -i "0.5"
+		_exit_code=$?
+	fi
+	return $_exit_code
+}
+
+function chroot_custom() {
+	local target=$1
+	shift
+	run_host_command_logged_raw chroot "${target}" /bin/bash -e -c "$*"
+}
+
+# for long-running, host-side expanded bash invocations.
+# the user gets a pv-based spinner based on the number of lines that flows to stdout (log messages).
+# the raw version is already redirect stderr to stdout, and we'll be running under do_with_logging,
+# so: _the stdout must flow_!!!
+function run_host_command_logged_long_running() {
+	local _exit_code=1
+	if [[ "${SHOW_LOG}" == "yes" ]] || [[ "${CI}" == "true" ]]; then
+		run_host_command_logged_raw /bin/bash -e -c "$*"
+		_exit_code=$?
+	else
+		run_host_command_logged_raw /bin/bash -e -c "$*" | pv -N "$(logging_echo_prefix_for_pv "${INDICATOR:-compile}")" --progress --timer --line-mode --force --cursor --delay-start 0 -i "0.5"
+		_exit_code=$?
+	fi
+	return $_exit_code
+}
+
+# run_host_command_logged is the very basic, should be used for everything, but, please use helpers above, this is very low-level.
+function run_host_command_logged() {
+	run_host_command_logged_raw /bin/bash -e -c "$*"
+}
+
+# do NOT use directly, it does NOT expand the way it should (through bash)
+function run_host_command_logged_raw() {
 	# Log the command to the current logfile, so it has context of what was run.
 	if [[ -f "${CURRENT_LOGFILE}" ]]; then
-		echo "(=-Armbian-->" chroot "\${SDCARD}" /bin/bash -e -c "$*" " <- at $(date --utc)" >> "${CURRENT_LOGFILE}" # SDCARD's $ is escaped on purpose here.
+		echo "       " >> "${CURRENT_LOGFILE}" # blank line for reader's benefit
+		echo "-->" "$*" " <- at $(date --utc)" >> "${CURRENT_LOGFILE}"
 	fi
+
+	# uncomment when desperate to understand what's going on
+	# echo "cmd about to run" "$@" >&2
+
 	local exit_code=666
-	chroot "${SDCARD}" /bin/bash -e -c "$*" 2>&1 # redirect stderr to stdout. $* is NOT $@!
+	"$@" 2>&1 # redirect stderr to stdout. $* is NOT $@!
 	exit_code=$?
 	if [[ -f "${CURRENT_LOGFILE}" ]]; then
-		echo "(=-Armbian--> cmd exited with code ${exit_code} at $(date --utc)" >> "${CURRENT_LOGFILE}" # SDCARD's $ is escaped on purpose here.
+		echo "--> cmd exited with code ${exit_code} at $(date --utc)" >> "${CURRENT_LOGFILE}"
+	fi
+	if [[ $exit_code != 0 ]]; then
+		display_alert "cmd exited with code ${exit_code}" "$*" "wrn"
 	fi
 	return $exit_code
 }
@@ -228,6 +341,5 @@ install_deb_chroot() {
 
 # @TODO: logging: used by desktop.sh exclusively. let's unify?
 run_on_sdcard() {
-	# Lack of quotes allows for redirections and pipes easily.
-	chroot "${SDCARD}" /bin/bash -c "${@}" >> "${DEST}/${LOG_SUBPATH}/install.log"
+	chroot_sdcard "${@}"
 }
